+++ docs/changelog.rst
@@ -35,6 +35,7 @@ Additions to the :doc:`primary modules <primary-modules>`.
     * :py:func:`freshpy.workspaces.get_workspace`
     * :py:func:`freshpy.workspaces.get_all_workspaces`
 * Added the :py:mod:`freshpy.objects` module with the following functions:
+    * :py:func:`freshpy.objects.get_custom_object`
     * :py:func:`freshpy.objects.get_all_custom_objects`
 * Added the :py:func:`freshpy.api.api_call_with_payload` function
 * Added the :py:func:`freshpy.api._raise_exception_for_api_method` function